+++ src/contents/ui/configGeneral.qml
@@ -19,40 +19,18 @@ import "../js/layouts.js" as Layouts
 Kirigami.FormLayout {
 	Layout.fillWidth: true
 
-	property alias cfg_layoutKey: layoutKey.text
+	property alias cfg_layoutKey: layoutSelector.selectedLayoutKey
 	property alias cfg_useUserLayout: useUserLayout.checked
 	property alias cfg_useSpecificLocaleEnabled: useSpecificLocaleEnabled.checked
 	property alias cfg_useSpecificLocaleLocaleName: useSpecificLocaleLocaleName.text
 	property alias cfg_useCustomFont: useCustomFont.checked
 	property alias cfg_customFont: fontSelector.selectedFont
 
-	Text {
-		visible: false
-		id: layoutKey
-	}
-
-	PlasmaComponents.ComboBox {
-		Kirigami.FormData.label: i18n('Layout')
+	LayoutSelector {
+		id: layoutSelector
 		enabled: !cfg_useUserLayout
-		textRole: "text"
-		model: []
-		Component.onCompleted: {
-			// populate model from Theme object
-			var tmp = []
-			var idx = 0
-			var currentIdx = undefined
-			for(const key in Layouts.layouts) {
-				var name = Layouts.layouts[key]['name']
-				tmp.push({'value':key, 'text': name})
-				if (key === plasmoid.configuration['layoutKey']) currentIdx = idx
-				idx++
-			}
-			model = tmp
-
-			currentIndex = currentIdx
-		}
-		onCurrentIndexChanged: cfg_layoutKey = model[currentIndex]['value']
-	} // ComboBox
+		Kirigami.FormData.label: i18n('Layout')
+	}
 
 	PlasmaComponents.CheckBox {
 		id: useUserLayout